+++ HexResourceTracker/Core/Tracking/TrackingModeEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace HexResourceTracker.Core.Tracking
+{
+    internal enum TrackingModeEnum
+    {
+        ZoneBased,
+        RangeScanner
+    }
+}