+++ HexResourceTracker/Core/TrackingRangeService.cs
@@ -1,0 +1,30 @@
+﻿using HexResourceTracker.Core.Tracking;
+using UnityEngine;
+
+namespace HexResourceTracker.Core
+{
+    internal static class TrackingRangeService
+    {
+        internal static bool IsWithinTrackingRange(Vector3 position)
+        {
+            if (PluginConfig.TrackingMode.Value == TrackingModeEnum.ZoneBased)
+            {
+                return true;
+            }
+
+            if (Player.m_localPlayer == null)
+            {
+                return false;
+            }
+
+            float range = PluginConfig.TrackingRange.Value;
+            Vector3 playerPosition = Player.m_localPlayer.transform.position;
+
+            float deltaX = playerPosition.x - position.x;
+            float deltaZ = playerPosition.z - position.z;
+            float distanceSqr = (deltaX * deltaX) + (deltaZ * deltaZ);
+
+            return distanceSqr <= range * range;
+        }
+    }
+}