+++ HexResourceTracker/Patches/PatchPickableAwake.cs
@@ -1,5 +1,6 @@
 ﻿using HarmonyLib;
 using HexResourceTracker.Core;
+using HexResourceTracker.Core.Tracking;
 
 namespace HexResourceTracker.Patches
 {
@@ -9,6 +10,18 @@ namespace HexResourceTracker.Patches
         private static void Postfix(Pickable __instance)
         {
             if (!PluginConfig.IsModEnabled.Value)
+            {
+                return;
+            }
+
+            string prefabName = __instance.gameObject.name.Replace("(Clone)", string.Empty).Trim();
+
+            if (PluginConfig.ResourceConfigs.ContainsKey(prefabName))
+            {
+                TrackedMapObject.TryAdd(__instance.gameObject, prefabName);
+            }
+
+            if (PluginConfig.TrackingMode.Value == TrackingModeEnum.RangeScanner)
             {
                 return;
             }